requirements.txt
asyncio==3.4.3
asyncpg==0.29.0
pydantic==2.7.1
python-dotenv==1.0.1
aiohttp==3.9.5
tenacity==8.2.3
structlog==24.1.0
typer==0.12.3
pytest==8.2.0
pytest-asyncio==0.23.7
alembic==1.13.1
sqlalchemy==2.0.30
